# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/avatar/chat/stream:
        post:
            tags:
                - AvatarService
            description: |-
                流式聊天（服务端以 SSE 持续返回 AI 回复片段；前端逐段渲染）
                 备注：本接口仅支持 HTTP，GRPC 下可另行定义 bidi-stream。
            operationId: AvatarService_ChatStream
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.avatar.v1.ChatRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.avatar.v1.ChatReply'
...
components:
    schemas:
...
        api.avatar.v1.ChatReply:
            type: object
            properties:
                messageId:
                    type: string
                    description: 本次用户消息的ID
                content:
                    type: string
                    description: 本次用户消息内容回显
                createdAt:
                    type: string
                    description: 发送时间 YYYY-MM-DD HH:MM:SS
                aiMessageId:
                    type: string
                    description: （可选）AI 回复的消息ID
                aiContent:
                    type: string
                    description: （可选）AI 回复的文本
                aiCreatedAt:
                    type: string
                    description: （可选）AI 回复时间 YYYY-MM-DD HH:MM:SS
        api.avatar.v1.ChatRequest:
            type: object
            properties:
                content:
                    type: string
                    description: 用户输入的文本内容
            description: 聊天
...